qcacld: Fix to configure gTxBFCsnValue based on the transmt side
Currently gTxBFCsnValue is based on the max value configured for the
STA. This does not take into consideration about the transmit side

This fix will consider the max antenna supported in the transmit
side as well before setting the gTxBFCsnValue.
WNI_CFG_VHT_CSN_BEAMFORMEE_ANT_SUPPORTED_STAMAX is also updated to 4.

Change-Id: If26f49386b72864730679e05559b7bba80b5487a
CRs-Fixed: 709938