wlan: qcacld: Channel Info in Roam Synch Indication
If DS Params and HT IE Info are not present in the
probe response, then the channel info would need to
be explicitly passed by the firmware to the host as
part of Roam Synch Indication. The information thus
retrieved from the firmware should be used to fill
the information in the BSS Descriptor.

Change-Id: I7f0caab1a52bf92d4c95290aa46bd33a977e7266
CRs-Fixed: 714313