wlan: qca-cld: SETROAMMODE Fix
When SETROAMMODE is used to disable LFR,
send the STOP command to firmware and then
configure the LFR2 ini param to Zero(disable).

When SETROAMMODE is used to enable LFR,
configure the LFR2 ini param to one(enable) and
then send the START command to firmware.

The roaming related parameters are passed during
the START command. So, there is no need to send
them seperately during the vdev up/down operations.

CRs-Fixed: 721194
Change-Id: I6300077369fd0356ecbc0f297b90c3ee472973b6