qcacld: Set connState to not connected, if disconnect cmd is not queued
The analysis of this issue is, if the function sme_RoamDisconnect returns
eHAL_STATUS_CMD_NOT_QUEUED (which means command not queued because there
is no connection). At this point state m/c is set to Disconnecting and
we are starting the disconnect timer which timesout and returns -EETIMEDOUT
and hence connState is not set to eConnectionState_NotConnected.

When the connection state "Connecting", it can move to either "Connected"
or "Disconnected". Here since there is a connection failure, we should be
moving to "Disconnected". Upon connection failure and hdd disconnect in
progress, we don't move to "Not connected" state and state is stuck at
"Connection in progress". Hence scan failure.

To fix this issue, if sme_RoamDisconnect returns eHAL_STATUS_CMD_NOT_QUEUED
move the state to "Not connected"

Change-Id: I174eda7eec679f5861946d04c78ad1e76a17975e
CRs-Fixed: 728064