qcacld: Fix sending a missing roam scan offload to firmware.
Currently, when a disconnect comes from the supplicant with reason
code as DEAUTH, the host is skipping doing a Link Down as a result
of which after a subsequent new connection the host misses sending
a roam scan offload to firmware because of being in a wrong state.
This issue is now fixed by calling link down before disconnecting.

Change-Id: I3a3926baca8a1324540e0db564430e59a5dd2fbd
CRs-Fixed: 728743