qcacld: ExtScan: Fix invalid timestamp in cached results
The expectation from the framework is to return the time since boot
(in microsecond) when the BSSID was cached.
Firmware returns the timestamp from WiFi turn ON till BSSID entry
was cached. Host driver shall add the time gap between DUT boot up to
WiFi turn ON to derive the time since boot when the BSSID was cached.

Change-Id: I087bbca1051bdd92bee9405a31ed6b30cce23cf6
CRs-Fixed: 732680