wlan: Dynamically enable/disable DFS channel scan.
All types of scans on DFS channels can be dynamically enabled
or disabled. By default, the passive scan on DFS channels is
enabled. When the scan is disabled, pass a new channel list to
the firmware that has the channels disabled. Also, do not honor
the dynamic command if SAP is already running on DFS channel.

Change-Id: I3fcf975f2cdf343ec481fbecb30dc1657d1a8350
CRs-Fixed: 729786