qcacld: cleanup the hdd_config_t config struct
Over time the HDD config struct has supported many configuration
items.  In some cases support for a configuration item was removed,
but the associated field in the config struct was not removed.  In
other cases a field for new configuration item was added in the middle
of a set of related fields.  Remove the unused fields and rearrange
the ones where a group of related fields were separated.

Change-Id: I8d3f57141d9490397df771b6a8c543434a9ed58b
CRs-fixed: 736787