qcacld: hdd: add condition to check to acquire dfs wakelock
When SAP operates in DFS channel, sap_dfs_ref_cnt is
used to keep tracking the wakelock that has been acquired
to prevent the system go into suspend.

When start the SAP, it'll acquire the lock first,
0. acquire lock for old channel A (sap_dfs_ref_cnt++)

And when radar comes in, the following logic will happen
1. acquire lock for new channel B (sap_dfs_ref_cnt++)
2. release lock for old channel A (sap_dfs_ref_cnt--)

Also the bssState will be used as sainty checking,
to decide if we've to acquire/release the lock or not.

With above logic, the bssState is already BSS_START.
So that the step#1 to acquire lock on new channel would never
be able to done. And only the step#2 release lock on old channel
will be done.

This then allow the system goes into suspend state.

Add the sap_df_ref_cnt as the combo condition to decide
if we have to acquire the lock or not.

Change-Id: Id2c565a2cd48c419Montana45b276cabbf8279ab9c47
CRs-fixed: 751325