qcacld: Fix the handling of invalid input to SETBAND
The root cause, if the command is "SETBAND 100" the existing code reads
only the 8th byte and ignores the following bytes and hence "SETBAND 100"
is treated similar to "SETBAND 1" and proceeds without throwing error.

The fix is to read the entire ASCII string and convert it to integer
and let the code take care of error handling.

Change-Id: Ibcd1a7a5c88e9b05c49d54ef3c7163be536c5f25
CRs-Fixed: 752065