qca-cld:sap: Fix ACS channel selection for HT40 and VHT80 cases
Currently the maximum weightage of a single channel is used as a threshold
to determine if a channel can be picked. But in HT40 and VHT80 case the
combined weightage of channels may exceed this threshold value. This fixes
the threshold values for determine if a channel should be picked.

Change-Id: I71ae3c48b329cca7989874034cfd9254c04520f5
CRs-fixed: 759593