qcacld: hdd/ipa lower the logging level
When running the throughput test, if certain peers get
disconnected for some reason, and before the peer is completely
removed.

This packet will be forward to HDD, and the message would keep
printing, which would hog the cpu resource until all the packets
in queue get drainned, we would see watchdog barks sometimes.

So lower the logging level to avoid this case.

Change-Id: I0b1c47a842bbc4aa1484aad5691a8c1704bc8f19
CRs-fixed: 759659