qcacld: [HDD] Add support for PCIe link auto suspend feature
When there is no active traffic to the firmware, bus can be
suspended. In the current design, bus is suspended only when
kernel suspends. However if driver can detect the inactivity
at the bus, it can suspend the bus autonomously when it is not
in use. HDD will start a timer to continously attempt bus suspend
when it is in STA mode (not connected).

CRs-Fixed: 761959
Change-Id: I6a036fb90d9235e9ec7495d3c4af1b7a7d6dfcbd