qcacld: Mask interrupts during driver load failure
During driver load failure, interrupts are not
masked as part of clean up. So, this leads to a
race condition where the memory freed during clean
up is accessed in the interrupt handler. This fix
ensures that the interrupts are masked and hence
no null pointer deference can happen for the freed
up pointers.

Change-Id: Ia859aa42a1af6ec3585a1e7137cddcca0377b5fc
CRs-Fixed: 776489