HDD: Update SAP operation ch on Ch Change Event
On RADAR detection SAP changes the operating channel and
SME/PE is already updated for new operating channel. So update
HDD layer also here. This resolves an issue in AP-AP mode where
AP1 channel is changed due to RADAR then CAC is going on and
START_BSS on new channel has not come to HDD. At this case if
AP2 is started it needs current operating channel of AP1 for
MCC DFS restriction.

Change-Id: I7a28232406a9bbb36e198bbbd9509a27f1801efc
CRs-fixed: 786915