qcacld-2.0: Failed to Send Action frame
This is prima to qcacld-2.0 propagation.

When GO exist and invites peer on the same operating channel where
GO is formed, i.e when GO is on channel, if supplicant sends,
a non zero wait time, driver is not able to send action frame.

As a part of fix, set wait time as zero when GO is operating on channel
explicitly in driver.

Change-Id: I608cf67692595658b79d4322c3d5a0705d2bee69
CRs-Fixed: 795809