qcacld: Do vdev restart during processing of change in HT params
In the curernt code, we do a vdev start through switch channel
whenever there is change in HT capabilities in beacons or probe
responses. But, vdev up is done only on receiving vdev restart
response event. Since, the current implementation only sends
vdev start, the host driver is not sending vdev up and hence
some of the de-registered offloads like CSA are not being
enabled back. To fix, this vdev start is being replaced by vdev
restart in the above scenario.

Change-Id: I05f774fecc79e17965b2fe43cfbd61c485c3ec5c
CRs-Fixed: 802970