qcacld: update cfg with valid channels and txpower values.
prima to qcacld propagation

With the current implementation updated channel list and tx
power values are not sent to firmware as cfg updation is
done after sending the values to the firmware. This commit
ensures cfg is updated with valid channel list and tx power
values before sending the same to firmware.

Change-Id: I00af7a1bd1df3235ce1769803c08033519b55308
CRs-Fixed: 808223