qca-cld: Set WMM enable/disable for WNI config
WNI parameter WNI_CFG_WME_ENABLED is used to determine if WMM IE
is present in beacon or not, however it is not being set based
on the INI parameter WmmIsEnabled at SAP start up time. This fix
initializes the WNI parameter value based on the INI config setting.

Change-Id: I1c36c7acbcbfe4f557dacc633807ac821891cac5
CRs-fixed: 811354