wlan: Avoid race condition in SAP mode
There is race condition seen when upper layer sends the deauth
to the STA and at same time STA sends the assocReq frame to SAP.
In this case LIM sends the Assoc indication to SME and wait for
the Assoc cnf. Now SME passes the deauth request to Lim and LIM
deletes this pStaD. When SME sends the assoc Cnf it checks pStaD
as NULL. So LIM and SME are not in sync here.Fix that.

Change-Id: Iee844eaec543b0a2f1f7ec9d9ef26f464df5697c
CRs-Fixed: 814751