qcacld: Reset the last radar found channel during P2P MHS GO restart
When P2P MHS is brought up on a channel on which radar was detected
previously, the last radar found channel is reinitialized. If this
reset is not done, the subsequent radars detected will not be
indicated to HDD since this will be treated as radars being
detected on the last detected channel again.

Change-Id: I611c8c3a583c49e67f56039f6d4de0e7fc4d8553
CRs-Fixed: 815061