qcacld-2.0: Fix the comeback timer value if its less than 10ms
When driver receives comeback time value as part of assoc response
less than 10ms then it fails to start the timer because of min
allowed timer value and due to this timer failure, driver doesn't
re-attempt the connection. As part of this patch, when driver
receives any value less than 10ms then it will start the timer
with 10ms.

Change-Id: If32c789084b09db6d00243d668917c9c79f3ef26
CRs-Fixed: 818055