qcacld: Cleanup saved Pre-Auth request on Pre-Auth failure
While doing a pre-auth, the pre-auth request for which memory is
dynamicaly allocated is cached in LIM and is not freed until a
subsequent roaming happens, but lim also relies on this request
being not null to allow a channel switch caused by a change in HT
parameters. As a result after one roaming is done, a HT parameters
change in beacon is not processed. This change ensures that this
scenario is fixed by doing the following-
1. Freeup the cached pre-auth request after roaming is done and
2. Cleanup FT context if pre-auth resulted in a failure.

Change-Id: Ic581b55ac7e08de84bdc9830747c4a921be20f3b
CRs-Fixed: 818207