qcacld: Fix OCB configuration issues
When configuring the channels for OCB mode, maximum TX
power from the regulatory domain is sent to the firmware
to be used for TX power control. Mac addresses no longer
leak when setting the OCB channel. The bandwidth parameter
has been added for setting the OCB channel via the wext
interface. The tx power when setting the OCB channel is
treated as 1/2 dB increments.

Change-Id: I9c91b999fca099de808f27bab63d838d5b04f9fb
CRs-Fixed: 822516