qcacld-2.0: Deregistering STA with correct STA ID
This is a prima to qcacld-2.0 propagation.

In IBSS, whenever we receive eCSR_ROAM_IBSS_LEAVE
indication, we assign sta id as 0 which is for MC/BC
packets and pass it to hdd_roamDeregisterSTA. Thus,
only sta id 0 gets deregistered while sta id for
UC frames remains as it is.

This fix will deregister station both for UC frames
as well as BC/MC frames.

Change-Id: I52541453d882f94d36bcab6ff2a95899a3f7bdb1
CRs-Fixed: 730801