qcacld-2.0: Changes in handling pmkid cache
prima to qcacld-2.0 propagation

A delete pmksa operation looks for a matching pmksa
in the pmkid cache and if found zeroes it out. Hence
after a delete pmksa holes will be created in the pmkid
cache. It is necessary to re-adjust these holes created
in the cache so that upon subsequent set pmksa these
holes will be overwritten instead of overwriting the
valid pmksa present in the cache.

Change-Id: I3e4bac2c5097b7ff6b86798e4c5291dbafc98d7e
CRs-Fixed: 848505