qcacld-2.0: Ini scan parameters not effective on DFS channels
prima to qcacld-2.0 propagation

Currently in limContinuePostChannelScan API, for DFS channels
dwell time is taken as WNI_CFG_PASSIVE_MAXIMUM_CHANNEL_TIME.
This parameter is always set as gPassiveMaxChannelTime.
As a part of fix, made sure that for DFS channel dwell time
is reset as per the configuration in ini file for connected
session case.

Change-Id: Id2256b1388f602e11f3373150c624175b75c19f8
CRs-Fixed: 729426