qcacld-2.0: Ignore security mismatch for HS2.0 association
prima to qcacld-2.0 propagation

After connection, driver process beacon  to detect any change in
AP's capabilities. Currently if privacy bit is set in beacon and
if rsn or wpa bit is not set, driver issues disconnect command.

In case of HS2.0, all security capabilities are part of vendor
specific information. Hence, both wpa and rsn bit is 0, though
privacy bit is set. Since driver considers it as a change in AP
security params, it issues disconnect. As a part of fix, make sure
that for OSEN capabilities disconnect is not issued.

Change-Id: I800bea0448256f663f61bccf6b0ec6b0f71d824b
CRs-Fixed: 796397