qcacld-2.0: Fix for ERP bit not resetting after legacy AP goes away
ERP and HT protection bits are set when a legacy AP (11b only) comes up
in vicinity of SAP. The same bits should be reset when all legacy APs are
gone. The reset function that gets called at regular interval was not
resetting these bits. This patch fixes that.

Change-Id: Iee0a464f64ea2c6394bd2f9e87ef1a6fd0c92ce1
CRs-Fixed: 885388