qcacld-2.0: Initial 2G only Scan fix
Current implementation of initial 2g only scan is broken
in following ways:
1) Scan Id recycles so, keeping FIRST_SCAN_ID macro as 0x0001 will
make scan that comes after recyle of 0xFFFF behave as initial scan
2) function csrScan2GOnyRequest() is broken as it assigns a local
variable to channel list and when scan command is released later
it tries to free a non accessible memory and crashes.

This patch fixes both of these issues.

Change-Id: I93c99df2198c6e0b78e1c879e7b6f5e55bb8f122
CRs-Fixed: 829292