qcacld-2.0: Initial Connection with Zero Channel list
Roam start command should be sent to firmware, despite
the number of channels in the roaming channel map is
Zero. There might be cases where DFS channel roaming
might be disabled and we are connected to a lone AP
and that happens to be on a DFS channel. Since DFS
channel roaming is not enabled, the channel list of
this AP is not populated and channel count would be
Zero. But, roaming should still start so that when we
move to a different location we could find AP's on other
channels.

CRs-Fixed: 898262
Change-Id: If8f6c4604289e08bdccb17d24f89b8cc8b2071ac