qcacld 2.0: Changes to avoid taking spinlock on whole function
Initially, race condition was happening as two kthreads were
running tlshim_mgmt_rx_process due to which spinlock protection was
taken. The spinlock was taken on the whole function which causes
crash as the function is capable of going to sleep and as spinlock
was taken on the whole function, we cannot sleep within that
function.

This fix takes spinlock only on the part of the code which is
actually prone to race condition and which will also not go to
sleep.

Change-Id: I47a54d6777413ee42a4f451d2a4925e3fd6e958a
CRs-Fixed: 920024