qcacld-2.0: Crash on Fail to runtime resume
When runtime resume failed we are asserting and returning failure
to runtime pm framework. Driver Bus layer remains in resuming state
and any communication to FW is blocked by host and resulting in
stability crash after reaching 1024 pending commands in control path.

Kernel message buffer is overflowed by this time and logs are lost.
Hence crash the APPS at the earliest to catch the issue.

Change-Id: I24f4c1a5246843b04f6fb26edcab4f89d628b03d
CRs-Fixed: 931717