qcacld-2.0: Fix reset variable abortScan
prima to qcacld-2.0 propagation

When abort scan is requested in between finish scan request and its
response, variable abortScan is set in limProcessAbortScanInd which
is not cleared even after the completion of current scan. Because of
this, next scan is aborted during limProcessInitScanRsp. Fix this by
clearing variable abortScan if it is set during limProcessFinishScanRsp.

Change-Id: If1f85bf82aff2e4f877d6c53a30d8c218ce541e4
CRs-Fixed: 882723