qcacld-2.0: Fix Race between MC thread and tx completion tasklet
Crash is observed when htt packet is accessed after it's freed.
This happened because of race between MC thread to push the packet
to FW and tasklet which frees the packet in the tx completion event
from FW.

Hence check the tag info and store in a local variable, before
sending the packet to FW, and use that to perform relavent action.

Change-Id: Ib0d0bc13e4d290b64cff8381e0e0423af54bd235
CRs-Fixed: 953173