qcacld-2.0: Make skb->next = NULL for last skb
Whenever we receive a skb list from network stack, we prepare a
separate list of valid skbs. Originally we are not making
skb->next = NULL for the last skb in the list of valid skbs due
to which if we need to drop the last skb from the skb list
received from network stack, the pointer to that skb will
still be a part of the valid skb list and when we pass this list
to lower layers, it will try to access this skb or free it
and so there will be NULL pointer dereference and the system
crashes.

So, to fix the issue, make skb->next = NULL for last skb in the
list of valid skbs.

Change-Id: I9564063f1b7af963f0a4c90a8bf261a2bd9e9fec
CRs-Fixed: 956950