qcacld-2.0: Register management frames during wlan re init
prima to qcacld-2.0 propagation

It could be possible that after SSR, pMac context is lost,
due to which the management frame callback will not be invoked
and none of the management frames will be processed,
resulting in scan and connection issues.

To address this, register management frame callback during wlan re init.

CRs-Fixed: 962187
Change-Id: I5429da9dbc33c7e2044d5a4daa1c8d3d7af241ee