qcacld-2.0: Map non-FCC Countries to proper reg id
prima to qcacld-2.0 propagation.

Some non FCC countries are mapped to FCC. When these country
are set, host sends regdomain as FCC to firmware.
Now if channel 9 is set with secondary channel offset above primary
(HT40+ mode), the firmware returns failure as this combination is
not allowed in FCC.

Map non-FCC countries to proper reg id in countryInfoTable.

Change-Id: I85313c1ce97e55b70dae1b28c91b4e703f5c7851
CRs-Fixed: 957368