ARM: dts: msm: Import msm8996 energy model
Add energy model for mtp8996

Change model to describe SMP for most of the power curve. The siver/gold
clusters are very close and may only differ by measurement error for
most of the curve. Several workloads (including video playback)
have better power characteristics if we go ahead and make the curves
identical so it looks like an SMP system for frequencies above the
min and below the silver max.

Change-Id: I5f8dcb11372b15549df7d0b03a200afd80db85d3