qcacld-2.0 Reduce LL stats / Firmware Memory Dump Blocking time
prima to qcacld-2.0 propagation

Currently driver block wpa_supplicant context for response
for 5 / 5 seconds for LL stats and firmware memory dump
feature respectively. But in M framework watch dog will
be triggered after 1 second because driver is holding
the context longer, which lead to wd thread starvation.
As a part of this fix, block time is reduced to 800 msec
respectively in both cases.

Change-Id: I02cc7623bc0bc6004ece972f8c4a5e8e751131d2
CRs-Fixed: 951119