qcacld-2.0: Update allowed action frames to FW for non STA mode
When the host driver operates in AP modes(SAP/P2P), FW should not drop
the HT action frames. It should send it to host for processing. Update
the FW with the allowed action frames (include HT action) in WOW mode
when the host starts SAP/GO session and reset when there are no SAP/GO
session.

Change-Id: Iae5420114ba7cfa2f98dc6c346d99d2f820c8503
CRs-Fixed: 1018376